+++ packages/style/src/helpers/isNestedSelector.ts
@@ -1,5 +1,6 @@
 const PATTERN = /^(:|\[|>|~|\+|\*)/u;
 
 export default function isNestedSelector(value: string): boolean {
+  // https://jsperf.com/string-startswith/66
   return PATTERN.test(value);
 }
